// src/enums/groupEnum.ts
export enum QueryGroupTypeEnum {
    BY_CREATE_DATE = 'byCreateDate',
    BY_START_DATE = 'byStartDate',
}

export enum GroupStatisticsSettingsEnum {
    STATISTIC_SHOW_COLUMNS = 'statisticShowColumns'
}

export enum GroupStatisticsColumnsEnum {
    PRODUCT_NAME = 'product_name',
    VENDOR_NAME = 'vendor_name',
    CREATE_GROUP_COUNT_BY_THE_MONTH = 'create_group_count_by_the_month',
    REVENUE_SUM = 'revenue_sum',
    ORDER_COUNT = 'order_count',
    CREATE_GROUP_COUNT_BY_MONTH = 'create_group_count_by_month',
    CREATE_GROUP_COUNT_BY_TOTAL = 'create_group_count_by_total',
}